Adds the query string to the search box when the user clicks either search button in the AdvancedSearchView #76
However, I'm getting more than a little bit worried about how we are going to
deal with all the different query strings that a user could put in.

Maybe if the map function was changed to return an array of all the datum
fields then I'd be able to perform the AND and OR operations when I get back.
I think I like this. Then it's only one view that doesn't need to be
modified no matter how many custom datafields there are.


Former-commit-id: 66940e63811f44cae22c6445fc2fe4c3d62020a3