The datum state dropdown is now being correctly populated and the selected datum state is being saved correctly. #249
When you create a duplicate datum, the datum state is also duplicated.
The default datum state is the first in the dropdown list.

No longer using a DatumStateEditView because we didn't use an updating
collection view for the dropdown menu like we do when we create lists
(like in the corpus). Instead, it is just done using the DatumEditView's
handlebars.


Former-commit-id: 3db61030dd45f2c5d5122559a9f3b2c213eced30