Fixed double click problem when creating new datum mentioned in #07fd672. For #249.
@mecathcart @cesine Take note of how I fixed the double click issue because
we're going to have to do the same thing for all of the items in the
dropdown menu.
For some reason, both the "click new_datum" event in CorpusEditView and
the "click new_datum" event in CorpusReadView were being triggered when I
clicked on "New Datum" in the dropdown menu. To fix it, I just went into
all the handlebars and renamed the class to new_datum_edit if I wanted to
trigger it in the CorpusEditView and renamed the class to new_datum_read if
I wanted to trigger it in the CorpusReadView.
This is maybe not the most elegant fix, but it works.


Former-commit-id: 73b2e3668273e7c67b405ca3e5739ac3a5f8fce9