Moved the datum session assignment outside of prependDatum. #322
This means that Datum.clone is now cloning the full datum, including the
session, which was not included in the method before.

This was done so that when I add a datum to the top of the stack, it isn't
necessarily put into the current session. This is applicable in the case that
a user double clicks on a datum in the data list from a previous session. It
should appear on top, but not be moved into the current session.


Former-commit-id: 527332f908cf8456520cacc643a01e8dcd5fecb3