fixes #508 here is a rough draft of importing elan
also added library so we can read any xml. fixes #506

How the import is working:
it takes each annotation of each tier as a datum, although the elan data is across tiers whcih means this should be re-written to use the elan xml specs and the cross referenceing with in them.

However, rather than perfecting the elan import as xml, we need to know if this is what users want. Do users want to import eaf? If they do, why do they want ot import eaf?

t would be far easier ot import one of ELANs other export formats like tabs, interlinear glosses etc. so we might simply reccomend that users dont import eaf, @balamkej @jessicacoon  this is something we will need input on from someone who actually has their database already in elan. they know elan best...

Our original motivation to use eaf is so that import and export to iField is non-lossy, so that users can use ifield to search, and gloss data, and continue to use ELAN as their primary user interface. What is the probability that real elan users will use ifield. I think it is very low, so this kind of non-lossy import export will take up a lot of development time with no one using it.

What would they use ifield for, if they used it. Would it be for search and glossing? ELAN is opensource and in java, we can write a webservice that does autoglosing for elan users, and they can keep using elan...

Alternatives:
* elan exported as interlinear gloss (see sample_elan.eaf)
* elan exported as praat textgrid (see sample_elan.TextGrid)

I will be using the TextGrid format as the results of the prosody lab aligner, it will be easy for iField to import and export TextGrid. Can TextGrid be our primary format for time aligned transcriptions... how well can elan import and expot textgrids?


Former-commit-id: 6df2292c38b3e2311adc29777d4d30ed0ef6252b