Added paragraph breaks and clarified text in corpus settings descriptions
Text is less wall-y now! Also tried to make instructions clearer and
more consistent with each other and with the app as it is now.

I don't think there is an "All Data" link, so I took that instruction
out. Instead I told them to search for a blank term.


Former-commit-id: 7b4b6138fb22014076a622c0c9b56c84324751ad