added  xhrFields: {withCredentials: true} everywhere in jquery couch
we now have an apparently CORS enabled jquery couch, or at least a rough draft. I can log in, open data, save data in the chrome app while contacting a couchdb wrappd in https://github.com/OpenSourceFieldlinguistics/CorpusWebService the CORS-proxy project customized for couchbd and nginx or another proxy.

also removed the reference to browser.msie

http://bugs.jquery.com/ticket/13214

so we could upgrade to jquery 1.9 fixes #849

http://stackoverflow.com/questions/11728343/set-cookie-session-not-being-set-after-sign-in-with-cross-domain

We were getting cookies fine, it was hard to tell but you can tell by opening another tab and getting into the databse and seeing you're logged as the intended user. The problem was that jquery couch wasnt sending credentials back so the couchdb couldnt tell that we were in the session. When not in a CORS situation the cookie is added but you must explicitly add it if you are in  CORS relationship to the database.

and apparently fixes #838 still more tests to confirm that this doesnt in turn break the couchapp when it runs out of the actual database.


Former-commit-id: 1a623156db768595b47638825893b138846fa6bf