fixes #907 and #906
added a new machine field called syntacticCategory which will be used for cleaning and search but not necessarily somethign that most users will use. it is hidden all of the time andonly visible if you click on the show rare fields button. also modified the new corpus logic to not copy the datum fields if you making a new corpus from the firstcorpus (a practice corpus where we expect people to have garbage that they dont want. later for #806 we will make a dropdown with a set of tempaltes that the user can choose...ie, language community, linguist, corpus linguist, phonetician etc


Former-commit-id: b942bbebe6eccdd1177609ecbf2de875caa2753b