Added new packages (database and contentprovider). Added new classes (sqlite helper, table class, content provider, and edit activity). Decided to move the editing/creating of records to their own activity. It's just SO much easier in the end vs. one fragment container that would hold not only the new/edit regular fragment, but also a webview fragment.
CloudDetailFragment needs to be able to change the dummy mItem variable to something real that will then be passable to the javascript so that the JS will be aware of the text from the sqlite db.

CloudEditActivity needs to be fixed so that not only will it create new records, but that it also takes the proper call to edit records.